requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.0.0

# LLM clients
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0

# NLP and text processing
nltk>=3.8.0

# Optional: For better performance
accelerate>=0.20.0
sentencepiece>=0.1.99
